Initial separation of custom OLI components and errors
